Factor out type converter so it could be overriden if necessary
